+++ CHANGELOG.md
@@ -7,7 +7,9 @@ Fixed:
 - `@mappersmith/test`: Ensure case sensitive header keys are matched properly
 
 Removed:
+
 - `@mappersmith/core`: Remove deprecated `setContext` and injection of `context` in middleware params.
+- `@mappersmith/core`: Remove `middlewares` alias; use `middleware` instead.
 
 Changed:
 
@@ -144,9 +146,13 @@ Refactored:
 ## 2.35.0
 
 Fixed:
+<<<<<<< HEAD
 
 - Respect `allowResourceHostOverride` configuration in middlewares #240
-- A successful middleware should no longer overwriting a previous middleware's `error` #230
+- # A successful middleware should no longer overwriting a previous middleware's `error` #230
+  - Respect `allowResourceHostOverride` configuration in middleware #240
+  - A successful middleware should no longer overwriting a previous middleware's `error` #230
+    > > > > > > > 07c80e5 (Remove middlewares alias in favor of middleware (#352))
 
 Added:
 
@@ -453,13 +459,23 @@ Deprecated:
 
 ## 2.0.1
 
+<<<<<<< HEAD
+
 - bugfix: Send `resourceName` and `resourceMethod` when running the test lib
 - Add flag (`mockRequest`) when executing the middlewares from the test lib
 
 ## 2.0.0
 
 - Expose `resourceName` and `resourceMethod` to middlewares
+- # Features from 2.0.0-rc1 to 2.0.0-rc7
+  - bugfix: Send `resourceName` and `resourceMethod` when running the test lib
+  - Add flag (`mockRequest`) when executing the middleware from the test lib
+
+## 2.0.0
+
+- Expose `resourceName` and `resourceMethod` to middleware
 - Features from 2.0.0-rc1 to 2.0.0-rc7
+  > > > > > > > 07c80e5 (Remove middlewares alias in favor of middleware (#352))
 
 ## 2.0.0-rc7
 
@@ -470,7 +486,11 @@ Deprecated:
 
 ## 2.0.0-rc6
 
-- bugfix: `MockClient` should use the same middlewares configured in the client #37
+<<<<<<< HEAD
+
+- # bugfix: `MockClient` should use the same middlewares configured in the client #37
+  - bugfix: `MockClient` should use the same middleware configured in the client #37
+    > > > > > > > 07c80e5 (Remove middlewares alias in favor of middleware (#352))
 
 ## 2.0.0-rc5